make datamodel package public and add doc comments
