+++ ldai/client.go
@@ -4,10 +4,11 @@ import (
 	"encoding/json"
 	"fmt"
 
+	"github.com/launchdarkly/go-server-sdk/ldai/datamodel"
+
 	"github.com/alexkappa/mustache"
 	"github.com/launchdarkly/go-sdk-common/v3/ldcontext"
 	"github.com/launchdarkly/go-sdk-common/v3/ldvalue"
-	"github.com/launchdarkly/go-server-sdk/ldai/internal/datamodel"
 	"github.com/launchdarkly/go-server-sdk/v7/interfaces"
 )
