+++ ldai/client_test.go
@@ -4,11 +4,12 @@ import (
 	"errors"
 	"testing"
 
+	"github.com/launchdarkly/go-server-sdk/ldai/datamodel"
+
 	"github.com/launchdarkly/go-sdk-common/v3/ldcontext"
 	"github.com/launchdarkly/go-sdk-common/v3/ldlog"
 	"github.com/launchdarkly/go-sdk-common/v3/ldlogtest"
 	"github.com/launchdarkly/go-sdk-common/v3/ldvalue"
-	"github.com/launchdarkly/go-server-sdk/ldai/internal/datamodel"
 	"github.com/launchdarkly/go-server-sdk/v7/interfaces"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
